refactor: Command와 Service 러프하게 분리
