+++ src/main/kotlin/subcommand/NewsSubCommand.kt
@@ -5,16 +5,12 @@ import common.readFromFile
 import domain.News
 import kotlinx.cli.ExperimentalCli
 import kotlinx.cli.Subcommand
+import service.NewsService
 
 @OptIn(ExperimentalCli::class)
 class NewsSubCommand : Subcommand("news", "News") {
     override fun execute() {
-
-        val resultNews = readFromFile<MutableList<News>>(FileName.NEWS_LIST.fileName)
-        for (news in resultNews) {
-            println("${news.no} : ${news.title}")
-            println("url: ${news.url}")
-            println()
-        }
+        val newsService = NewsService()
+        newsService.getNews()
     }
 }
